add service and test

+++ field-audit/src/main/java/com/jianla/faudit/service/impl/QuestionServiceImpl.java
@@ -2,15 +2,19 @@ package com.jianla.faudit.service.impl;
 
 import com.jianla.faudit.dao.OptionDao;
 import com.jianla.faudit.dao.QuestionDao;
+import com.jianla.faudit.dto.QuestionDto;
 import com.jianla.faudit.entity.Option;
 import com.jianla.faudit.entity.Question;
 import com.jianla.faudit.service.QuestionService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -28,37 +32,40 @@ public class QuestionServiceImpl implements QuestionService {
     @Autowired
     private OptionDao optionDao;
 
-    @Override
-    public void create(Question question) {
-        questionDao.save(question);
-    }
 
     @Override
-    public void create(Question question, List<Option> options) {
+    public void create(Long orgId,String content, Short type, List<String> options) {
+        Question question = new Question(content,type,new Date(),orgId);
         questionDao.save(question);
         Long questionId = question.getId();
         logger.debug("create question[{}] success.",questionId);
-        for (Option option : options) {
-            option.setQuestionId(questionId);
+        for (String optionStr : options) {
+            Option option = new Option(optionStr,questionId);
             optionDao.save(option);
         }
     }
 
     @Override
-    public void update(Question question) {
-        questionDao.update(question);
+    public void update(Long id, String content, Short type, List<String> options) {
+        Question question = questionDao.getById(id);
+        if(question==null) return;
+        question.setContent(content);
+        question.setType(type);
+        //delete old options
+        Long questionId = question.getId();
+        optionDao.deleteByQuestionid(questionId);
+        //insert new options
+        for (String optionStr : options) {
+            Option option = new Option(optionStr,questionId);
+            optionDao.save(option);
+        }
     }
 
     @Override
-    public void update(Question question, List<Option> options) {
-        questionDao.update(question);
-        Long questionId = question.getId();
-        //delete all old options
-        optionDao.deleteByQuestionid(questionId);
-        //insert new options
-        for (Option option : options) {
-            option.setQuestionId(questionId);
-            optionDao.save(option);
+    public void delete(Long id) {
+        Question question = questionDao.getById(id);
+        if(question!=null){
+            this.delete(question);
         }
     }
 
@@ -71,12 +78,34 @@ public class QuestionServiceImpl implements QuestionService {
 
     @Override
     public Question getById(Long id) {
-        Question question = questionDao.get(Question.class, id);
+        Question question = questionDao.getById(id);
         return question;
+    }
+
+    @Override
+    public QuestionDto getDetailById(Long id) {
+        Question question = questionDao.getById(id);
+        if(question==null) return null;
+        QuestionDto dto = new QuestionDto();
+        BeanUtils.copyProperties(question,dto);
+        dto.setType(question.getType());
+        //find options
+        List<String> optionContents = optionDao.findContentByQuestionId(question.getId());
+        dto.setOptions(optionContents);
+        return dto;
     }
 
     @Override
     public List<Question> findByOrgId(Long orgId) {
         return questionDao.find("from Question where orgId = ?",new Object[]{orgId});
+    }
+
+    @Override
+    public List<QuestionDto> findDetailByIds(List<Long> questionIds) {
+        List<QuestionDto> list = new ArrayList<>();
+        for (Long questionId : questionIds) {
+            list.add(this.getDetailById(questionId));
+        }
+        return list;
     }
 }
